♻️ Refactor bot to support group handlers and remove unused public handlers
- Moved public chat handlers into a new grouphandlers package to better reflect their purpose in managing group chat interactions.
- Updated imports and handler registrations in `bot.go` to align with the changes.
- Removed the `publichandlers/save.go` file as it was no longer required.
- Introduced a constant for the cancel command in `adminhandlers/club_call_edit.go` for reusability and clarity.
- Enhanced the club call edit handler to validate private chat context and improve messaging when canceling an operation.
