+++ internal/handlers/grouphandlers/replies_from_closed_threads.go
@@ -1,4 +1,4 @@
-package publichandlers
+package grouphandlers
 
 import (
 	"fmt"
